Second Commit

+++ src/main/java/net/engineeringdigest/journalApp/service/JournalEntryService.java
@@ -1,23 +1,30 @@
 package net.engineeringdigest.journalApp.service;
 
+
+
+import lombok.extern.slf4j.Slf4j;
 import net.engineeringdigest.journalApp.entity.JournalEntry;
 import net.engineeringdigest.journalApp.entity.UserEntity;
 import net.engineeringdigest.journalApp.repository.JournalEntryRepo;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
-@Component
+@Service
+@Slf4j
 public class JournalEntryService {
     @Autowired
     private JournalEntryRepo journalEntryRepository;
     @Autowired
     private UserService userService;
+
+
 
     @Transactional
     public void SaveEntry(JournalEntry myje, String userName)
@@ -30,7 +37,7 @@ public class JournalEntryService {
             userService.saveUser(user);
         }catch (Exception e)
         {
-            System.out.println(e);
+
             throw new RuntimeException("An error occured while saving the entry.",e);
         }
 
@@ -50,11 +57,24 @@ public class JournalEntryService {
         return journalEntryRepository.findById(String.valueOf(id));
     }
 
-    public void DeleteById(ObjectId id, String userName)
+    @Transactional
+    public boolean DeleteById(ObjectId id, String userName)
     {
-        UserEntity user = userService.findByUserName(userName);
-        userService.saveNewUser(user);
-        user.getJournalEntries().removeIf(x-> x.getId().equals(id));
-        journalEntryRepository.deleteById(String.valueOf(id));
+        boolean flag = false;
+        try {
+            UserEntity user = userService.findByUserName(userName);
+            flag = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (flag) {
+                userService.saveUser(user);
+
+                journalEntryRepository.deleteById(String.valueOf(id));
+            }
+        } catch (Exception e) {
+            log.error("Error ",e);
+            throw new RuntimeException("Error occured while deleting the entry",e);
+        }
+        return flag;
+
+
     }
 }
